Running into issues getting tests to reference the source.

+++ src/Sampling/QualitySample.h
@@ -5,25 +5,25 @@
 
 class QualitySample
 {
-  public:
-    int CO2;
-    int TVOC;
-    int NOX;
-    int PM1;
-    int PM25;
-    int PM10;
-    float TemperatureC;
-    int RelativeHumidity;
+    public:
+        int CO2;
+        int TVOC;
+        int NOX;
+        int PM1;
+        float PM25;
+        int PM10;
+        float TemperatureC;
+        int RelativeHumidity;
 
-    QualitySample();
-    float TemperatureF();
-    int GetAQI();
+        QualitySample();
+        float TemperatureF();
+        int GetAQI();
 
-    static inline const String CO2Units = "ppm";
-    static inline const String ParticualteMatterUnits = "μg/m3";
-    static inline const String RelativeHumidityUnits = "%";
-    static inline const String TemperatureCUnits = "°C";
-    static inline const String TemperatureFUnits = "°F";
+        static inline const String CO2Units = "ppm";
+        static inline const String ParticualteMatterUnits = "μg/m3";
+        static inline const String RelativeHumidityUnits = "%";
+        static inline const String TemperatureCUnits = "°C";
+        static inline const String TemperatureFUnits = "°F";
 };
 
 #endif
